Port code to Clarisse 5.0 API
CSK package version updated to 2.0.
This CSK version is no longer compatible with previous Clarisse version (3.5, 3.6, 4.0) due to API changes in Clarisse 5.0. Shelf version list has been updated accordingly to patch only the 5.0 shelf.cfg.
Attribute "quality" in TextureOcclusion has been renamed to "sample_count".
OfAttr.get_type_name: this method is static in the class OfAttr, so it is recommended to call it using the syntax "ix.api.OfAttr.get_type_name". Especially now that OfAttr are mostly presented as OfAttrHandle (so it's not the same class).